requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.1
djangorestframework==3.12.2
env==0.1.0
environ==1.0
filelock==3.0.12
Pillow==8.0.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.30
